Changed to fix a compiler warning. Original: 24202eecb033e7df10f3c03810410e7f3fe22357

+++ lal/doc/laldoc/LALErrTableWriter.c
@@ -176,7 +176,7 @@ ParseErrLine( char *Ptr          , /* Ptr to string after #define)      */
     sscanf(position,"%s",errStr);
 
     /* Read the Error Number, ... */
-    if ( isdigit( errStr[0] ) ) {      
+    if ( isdigit( (int) errStr[0] ) ) {      
          sscanf(position,"%s",errNum);
          *errStr = '\0' ;
     }
